requirements.txt
aiocache==0.12.3
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
asgiref==3.8.1
asyncio==3.4.3
attrs==24.2.0
bcrypt==4.2.0
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
distlib==0.3.8
filelock==3.16.0
Flask==3.0.3
Flask-Cors==5.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
force==0.1.4
frozenlist==1.4.1
geographiclib==2.0
geopy==2.4.1
googlemaps==4.10.0
GoogleMapsApi==0.2.0
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
Hypercorn==0.17.3
hyperframe==6.0.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
MarkupSafe==2.1.5
memory-profiler==0.61.0
multidict==6.1.0
numpy==2.1.1
packaging==24.1
platformdirs==4.3.1
priority==2.0.0
propcache==0.2.0
psutil==6.1.0
psycopg2-binary==2.9.9
python-decouple==3.8
python-dotenv==1.0.1
RapidFuzz==3.10.0
requests==2.32.3
scikit-learn==1.5.2
scipy==1.14.1
SQLAlchemy==2.0.35
threadpoolctl==3.5.0
typing_extensions==4.12.2
urllib3==2.2.3
virtualenv==20.26.4
Werkzeug==3.0.4
wsproto==1.2.0
WTForms==3.1.2
yarl==1.15.2
